added shared util
